fix: include prev_proposer_pubkey in payload_id generation (#72)
This PR fixes payload_id generation to include the optional
prev_proposer_pubkey field in the hash calculation. The current
implementation was using the standard Ethereum payload_id function which
does not account for the Berachain-specific prev_proposer_pubkey field,
potentially causing payload conflicts when different proposer pubkeys
are used for the same block parameters in Prague1 scenarios. This
scenario is highly unlikely

NOTE: This does not get parity to the geth implementation as geth has
different behaviour for nil withdrawals and also included a payload
version byte, which reth doesn't have.

Fixes #68